sf_test_tool/requirements.txt
streamlit>=1.32.0
simple-salesforce>=1.12.5
litellm>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
playwright>=1.42.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.20.0
google-generativeai>=0.4.0
weasyprint>=60.0
jinja2>=3.1.0
Pillow>=10.0.0
xlsxwriter>=3.1.0
streamlit-aggrid>=0.3.4